pyproject.toml
[project]
name = "clawpm"
version = "0.1.0"
description = "Filesystem-first multi-project manager for AI agents"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "malphas", email = "malphas-gh@users.noreply.github.com" }]
keywords = ["cli", "project-management", "ai", "agents"]

dependencies = [
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "tomli>=2.0.0;python_version<'3.11'",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
]

[project.scripts]
clawpm = "clawpm.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/clawpm"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]
